requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-appconf==1.0.5
django-crispy-forms==1.14.0
django-etc==1.3.1

django-imagekit==4.1.0


oauthlib==3.2.0
pilkit==2.0
Pillow==9.0.1
py-vapid==1.8.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.1

requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0

sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
